1. Перенес модели в common. 2. Добавил очистку кеша после сохранения фильтров d sharedPref. 3. Внес изменения в SarchRepository, для того чтобы проект собирался
