+++ app/src/main/java/ru/practicum/android/diploma/common/domain/model/filter_models/Area.kt
@@ -1,0 +1,6 @@
+package ru.practicum.android.diploma.common.domain.model.filter_models
+
+data class Area(
+    var id: String,
+    var name: String
+)
