+++ app/src/main/java/ru/practicum/android/diploma/filter/data/dataSourceImpl/FilterOptionsDataSourceImpl.kt
@@ -1,12 +1,12 @@
 package ru.practicum.android.diploma.filter.data.dataSourceImpl
 
+import ru.practicum.android.diploma.common.domain.model.filter_models.Area
+import ru.practicum.android.diploma.common.domain.model.filter_models.Country
+import ru.practicum.android.diploma.common.domain.model.filter_models.Filter
+import ru.practicum.android.diploma.common.domain.model.filter_models.Industry
 import ru.practicum.android.diploma.filter.data.dataSource.FilterOptionsDataSource
 import ru.practicum.android.diploma.filter.data.db.FilterDataBase
 import ru.practicum.android.diploma.filter.data.db.FilterLocalCache
-import ru.practicum.android.diploma.filter.data.model.dto.Area
-import ru.practicum.android.diploma.filter.data.model.dto.Country
-import ru.practicum.android.diploma.filter.data.model.dto.Filter
-import ru.practicum.android.diploma.filter.data.model.dto.Industry
 
 class FilterOptionsDataSourceImpl(
     private val filterDataBase: FilterDataBase,
@@ -18,11 +18,11 @@ class FilterOptionsDataSourceImpl(
         } else {
             filterLocalCache.getFilterCache()
         }
-    filterLocalCache.clear()
     }
 
     override fun putFilterOptions(options: Filter) {
         filterDataBase.putFilterOptions(options)
+        filterLocalCache.clear()
     }
 
     override fun addCountry(country: Country) {
