+++ app/src/main/java/ru/practicum/android/diploma/filter/data/db/FilterDataBase.kt
@@ -1,9 +1,9 @@
 package ru.practicum.android.diploma.filter.data.db
 
-import ru.practicum.android.diploma.filter.data.model.dto.Filter
+import ru.practicum.android.diploma.common.domain.model.filter_models.Filter
 
 interface FilterDataBase {
     fun getFilterOptions(): Filter?
-    fun putFilterOptions(filter:Filter)
+    fun putFilterOptions(filter: Filter)
     fun clearSavedFilter()
 }
