+++ app/src/main/java/ru/practicum/android/diploma/filter/data/db/FilterLocalCache.kt
@@ -1,9 +1,9 @@
 package ru.practicum.android.diploma.filter.data.db
 
-import ru.practicum.android.diploma.filter.data.model.dto.Area
-import ru.practicum.android.diploma.filter.data.model.dto.Country
-import ru.practicum.android.diploma.filter.data.model.dto.Filter
-import ru.practicum.android.diploma.filter.data.model.dto.Industry
+import ru.practicum.android.diploma.common.domain.model.filter_models.Area
+import ru.practicum.android.diploma.common.domain.model.filter_models.Country
+import ru.practicum.android.diploma.common.domain.model.filter_models.Filter
+import ru.practicum.android.diploma.common.domain.model.filter_models.Industry
 
 interface FilterLocalCache {
     fun getFilterCache(): Filter?
