--- app/src/main/java/ru/practicum/android/diploma/filter/data/model/dto/Area.kt
@@ -1,6 +1,0 @@
-package ru.practicum.android.diploma.filter.data.model.dto
-
-data class Area(
-    var id: String,
-    var name: String
-)
