--- app/src/main/java/ru/practicum/android/diploma/filter/data/model/dto/Country.kt
@@ -1,6 +1,0 @@
-package ru.practicum.android.diploma.filter.data.model.dto
-
-data class Country(
-    var id: String,
-    var name: String
-)
